Migrate to glMultiDrawArraysIndirect

+++ src/main/java/edu/kit/scc/git/ggd/voxelite/render/ChunkShadowProgram.java
@@ -1,7 +1,6 @@
 package edu.kit.scc.git.ggd.voxelite.render;
 
 import net.durchholz.beacon.math.Matrix4f;
-import net.durchholz.beacon.math.Vec3f;
 import net.durchholz.beacon.math.Vec3i;
 import net.durchholz.beacon.render.opengl.OpenGL;
 import net.durchholz.beacon.render.opengl.shader.Program;
@@ -14,7 +13,7 @@ public class ChunkShadowProgram extends Program {
         super(shaders);
     }
 
-    public final Attribute<Vec3f>   position = attribute("pos", OpenGL.Type.FLOAT, 3);
+    public final Attribute<Vec3i>   position = attribute("pos", OpenGL.Type.INT, 3);
     public final Attribute<Integer> data     = attribute("data", OpenGL.Type.INT, 1);
 
     public final Uniform<Matrix4f> mvp            = uniMatrix4f("mvp", true);
